Feature/Create Task (#16)
* Create migration for our Task model

We want to persist named tasks so we create the appropriate table and
columns.

* Create Task model

We leave that model empty for now.

* Implement basic create method for tasks

We implement the bare minimum for task creation. Validation will be
implemented in another commit.

* Extract params into a private method

This way we can isolate which keys we are requiring/permitting.

* Return correct response on task creation

We inspect the data integrity in the returned response's body and check
that the status code is valid.

* Add validation for task names

We want all tasks to have a name so we add validation in the model.

* Clean up task spec

We inline small before blocks and rewrite the context strings.