Feature/Update Task (#18)
* Update task name on PATCH request

We want our user to be able to update task's names so we implement an
endpoint for that. We'll implement responses in a future commit.

* Return augmented task in the response

We want all of our endpoints to return the corresponding object on
success so we return the task and make sure it's updated.

* Return errors when task update fails

This method is starting to get big, we'll extract the logic in a future
commit.