Refactor/Early Returns Should Render Errors (#21)
* Extract error rendering into a private method

We remove some code repetition by moving the error rendering logic into
a method. We might create some helper object in the future to hold the
error logic outside of the controller.

* Return early on error in TaskController

It's unusual to return early on success when using guard statements so
we do the opposite and return only when an error happens.

* Move errors method under render_422

Following the newspaper metaphor, we move the method dealing with finer
details at the bottom under the caller `render_422`.