Feature/Delete Task (#23)
* Implement base for TasksController.destroy

We handle cases where the task is not found for the id. We also realised
we're not testing this with other actions like update. We'll come back
in a future commit and implement 404 for these.

* Raise error when task doesn't exist on update

After covering that case for the delete request, we come back to
patching tasks and implement it there also.

* Remove repetition when error rescuing

We use the `rescue_from` handler to remove rescuing from missing record
in every actions.

* Use path in request descriptions

We think using the path here is more insightful for the reader than the
internal controller method.