Fix/Uniformize Errors (#25)
* Add message key to 404 errors

In an effort to make our errors consistent and resilient to future
feature, we nest the error message under an error key.

* Conform 422 error with the 404 structure

We want all of our error to respect the same schema so we take the first
record issue and nest it under the message key.

* Move error rendering helper to base controller

These will be used in all of our controllers so we move them to the
ApplicationController base class. We might eventually extract these into
an helper object that we could unit test.