Refactoring/Errors (#28)
* Remove custom 404 error

We are now relying on the framework error instead of providing a very
similar custom implementation.

* Conform 422 to rails error schema

We want all of our errors to have the same structure so we use the rails
errors schema for our 422. We also add all of the record's errors to the
error body instead of only the first one, to provide more information on
what went wrong.