Record Task Completion (#30)
* Add completion date to Tasks table

We want to know when a task is completed, if the completion date is nil
then the task is still ongoing.

* Allow requests to update task's completion date

We give our api clients the ability to update a task's completion date
to classify it as done. Eventually we might expose a simple completed
boolean query param to set the task.completed_at to DateTime.now.