Filter Tasks based on State (#31)
* Fetch completed tasks

We want our user to be able to get only completed tasks if they want. To
do so we accept a state query param that maps to a task scope.

* Fetch open tasks

We give the api user ability to get only tasks that are not done.

* Create factory for open/completed tasks

This way we remove setup duplication and hide some implementation from
the request spec.

* Remove duplication from state scopes test setup

We move "state scopes" under a single describe to remove the setup
repetition.